requirements.txt
streamlit
matplotlib
numpy
Pillow
tensorflow
opencv-python